`qget` is _not_ safe. Fix it to be more data-safe and less program-safe
(Previous, if you interrupted it part way through it would not save
data like `chapter.processed` but would have downloaded updated pages,
meaning it wouldn't reprocess as it should because next time it would
consider it unchanged. This makes it save just prior to any Error,
or Interrupt, and then bubble it up. Which is hopefully not too
horrific. Oops. Only does it if something has changed.)